modularize onboarding component
